Fininshed the triangle angle solver function (Aiden, Ben, and Rhys did a live share to work colaboratively)

+++ PolygonChecker/main.h
@@ -1,4 +1,4 @@
 #pragma once
 void printWelcome();
 int printShapeMenu();
-int* getTriangleSides(int* triangleSides);
+double* getTriangleSides(double* triangleSides);
